backend/requirements.txt
Flask
langchain
google-generativeai
mysql-connector-python
python-dotenv
